requirements.txt
alembic==1.7.7
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
beautifulsoup4==4.11.1
bidict==0.22.0
certifi==2021.10.8
cffi==1.15.0
click==8.1.2
cryptography==36.0.2
Deprecated==1.2.13
Flask==2.1.1
Flask-Migrate==3.1.0
Flask-Script==2.0.6
Flask-SocketIO==5.1.1
Flask-SQLAlchemy==2.5.1
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.2
h11==0.13.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
lxml==4.8.0
Mako==1.2.0
MarkupSafe==2.1.1
outcome==1.1.0
packaging==21.3
psycopg2-binary==2.9.3
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.8
PySocks==1.7.1
python-dotenv==0.20.0
python-engineio==4.3.1
python-socketio==5.5.2
redis==4.2.2
rq==1.10.1
selenium==4.1.3
simple-websocket==0.5.2
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.35
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.9
Werkzeug==2.1.1
wrapt==1.14.0
wsproto==1.1.0
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
